requirements.txt
requests==2.31.0
ddddocr==1.4.11
cairosvg==2.7.1
Pillow==10.0.1
zai-sdk==0.0.4
